# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/dev_app/{device_id}/{amperage_application_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет параметры устройства в конкретной заявке",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "amperage_application_devices"
                ],
                "summary": "Изменить данные устройства в заявке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID устройства",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "amperage_application_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.AmperageApplicationDeviceJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленные данные",
                        "schema": {
                            "$ref": "#/definitions/serializer.AmperageApplicationDeviceJSON"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет связь устройства и заявки",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "amperage_application_devices"
                ],
                "summary": "Удалить устройство из заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID устройства",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "amperage_application_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная заявка",
                        "schema": {
                            "$ref": "#/definitions/serializer.AmperageApplicationJSON"
                        }
                    },
                    "400": {
                        "description": "Неверные ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "serializer.AmperageApplicationDeviceJSON": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "amperage": {
                    "type": "number"
                },
                "amperage_application_id": {
                    "type": "integer"
                },
                "app_dev_id": {
                    "type": "integer"
                },
                "device_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "serializer.AmperageApplicationJSON": {
            "type": "object",
            "properties": {
                "amperage": {
                    "type": "number"
                },
                "amperage_application_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_login": {
                    "type": "string"
                },
                "finish_date": {
                    "type": "string"
                },
                "form_date": {
                    "type": "string"
                },
                "moderator_login": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}